+++ console/src/main.ts
@@ -6,8 +6,6 @@ import { createRouter, createWebHashHistory } from 'vue-router'
 
 import 'vuetify/styles'
 import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
 
 import App from './App.vue'
 import Toast from './toast'
@@ -40,8 +38,6 @@ const router = createRouter({
 })
 
 const vuetify = createVuetify({
-  components,
-  directives,
   icons: {
     defaultSet: 'mdi'
   }