+++ console/src/tree/selection.ts
@@ -1,10 +1,13 @@
 export class Selector {
+  selected: HTMLElement | null
+  onSelect: ((node: HTMLElement, path: string[], isParent: boolean) => void) | null
+
   constructor() {
     this.selected = null
     this.onSelect = null
   }
 
-  select(node, path, isParent) {
+  select(node: HTMLElement, path: string[], isParent: boolean) {
     if (this.selected) {
       if (this.selected === node) {
         return false